Backend/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
Pillow>=10.0.0 
deepface
flask
flask-cors
tf-keras
psutil>=5.9.0